requirements.txt
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.8.2
bs4==0.0.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cryptography==45.0.5
cycler==0.12.1
DateTime==5.5
Deprecated==1.2.18
distro==1.9.0
dnspython==2.6.1
eventlet==0.39.1
exceptiongroup==1.3.0
finnhub-python==2.4.24
finvizfinance==0.14.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.8.0
Flask-SocketIO==5.5.1
flask-talisman==1.1.0
fonttools==4.57.0
greenlet==3.1.1
groq==0.30.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.3.2
kiwisolver==1.4.7
lightgbm==4.6.0
limits==3.13.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
mdurl==0.1.2
moomoo-api==9.3.5308
numpy==1.24.3
ordered-set==4.1.0
packaging==24.2
pandas==2.0.3
pandas_ta==0.3.14b0
pillow==10.4.0
polygon-api-client==1.14.6
protobuf==3.20.3
psutil==7.0.0
py-healthcheck==1.10.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
PyJWT==2.9.0
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
redis==6.1.1
requests==2.32.4
rich==13.9.4
schedule==1.2.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streaming_indicators==0.1.8
supervisor==4.2.5
threadpoolctl==3.5.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.6
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.0.1
zipp==3.20.2
zope.interface==7.2
